Minor updates for style and to minimize additions to the public API.
