--- api/src/main/java/org/apache/iceberg/WrappedPositionAccessor.java
@@ -1,46 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.iceberg;
-
-import org.apache.iceberg.types.Type;
-
-class WrappedPositionAccessor implements Accessor<StructLike> {
-  private final int position;
-  private final Accessor<StructLike> accessor;
-
-  WrappedPositionAccessor(int pos, Accessor<StructLike> accessor) {
-    this.position = pos;
-    this.accessor = accessor;
-  }
-
-  @Override
-  public Object get(StructLike row) {
-    StructLike inner = row.get(position, StructLike.class);
-    if (inner != null) {
-      return accessor.get(inner);
-    }
-    return null;
-  }
-
-  @Override
-  public Type type() {
-    return accessor.type();
-  }
-}
