+++ api/src/main/java/org/apache/iceberg/expressions/InclusiveManifestEvaluator.java
@@ -21,14 +21,11 @@ package org.apache.iceberg.expressions;
 
 import java.nio.ByteBuffer;
 import java.util.List;
-import org.apache.iceberg.Accessor;
+import org.apache.iceberg.Accessors;
 import org.apache.iceberg.ManifestFile;
 import org.apache.iceberg.ManifestFile.PartitionFieldSummary;
 import org.apache.iceberg.PartitionSpec;
-import org.apache.iceberg.PositionAccessor;
 import org.apache.iceberg.Schema;
-import org.apache.iceberg.StructLike;
-import org.apache.iceberg.exceptions.ValidationException;
 import org.apache.iceberg.expressions.ExpressionVisitors.BoundExpressionVisitor;
 import org.apache.iceberg.types.Conversions;
 import org.apache.iceberg.types.Types.StructType;
@@ -96,16 +93,6 @@ public class InclusiveManifestEvaluator {
       return ExpressionVisitors.visit(expr, this);
     }
 
-    private PositionAccessor getPositionAccessorForFieldId(int fieldId) {
-      Accessor<StructLike> accessor = schema.getAccessorForField(fieldId);
-      if (!(accessor instanceof PositionAccessor)) {
-        throw new ValidationException("Cannot recognize nested field for partition evaluation");
-      }
-      PositionAccessor positionAccessor = (PositionAccessor) accessor;
-
-      return positionAccessor;
-    }
-
     @Override
     public Boolean alwaysTrue() {
       return ROWS_MIGHT_MATCH; // all rows match
@@ -133,10 +120,10 @@ public class InclusiveManifestEvaluator {
 
     @Override
     public <T> Boolean isNull(BoundReference<T> ref) {
-      PositionAccessor positionAccessor = getPositionAccessorForFieldId(ref.fieldId());
+      int pos = Accessors.toPosition(ref.accessor());
       // no need to check whether the field is required because binding evaluates that case
       // if the column has no null values, the expression cannot match
-      if (!stats.get(positionAccessor.position()).containsNull()) {
+      if (!stats.get(pos).containsNull()) {
         return ROWS_CANNOT_MATCH;
       }
 
@@ -145,10 +132,10 @@ public class InclusiveManifestEvaluator {
 
     @Override
     public <T> Boolean notNull(BoundReference<T> ref) {
-      PositionAccessor positionAccessor = getPositionAccessorForFieldId(ref.fieldId());
+      int pos = Accessors.toPosition(ref.accessor());
       // containsNull encodes whether at least one partition value is null, lowerBound is null if
       // all partition values are null.
-      ByteBuffer lowerBound = stats.get(positionAccessor.position()).lowerBound();
+      ByteBuffer lowerBound = stats.get(pos).lowerBound();
       if (lowerBound == null) {
         return ROWS_CANNOT_MATCH; // all values are null
       }
@@ -158,8 +145,8 @@ public class InclusiveManifestEvaluator {
 
     @Override
     public <T> Boolean lt(BoundReference<T> ref, Literal<T> lit) {
-      PositionAccessor positionAccessor = getPositionAccessorForFieldId(ref.fieldId());
-      ByteBuffer lowerBound = stats.get(positionAccessor.position()).lowerBound();
+      int pos = Accessors.toPosition(ref.accessor());
+      ByteBuffer lowerBound = stats.get(pos).lowerBound();
       if (lowerBound == null) {
         return ROWS_CANNOT_MATCH; // values are all null
       }
@@ -176,8 +163,8 @@ public class InclusiveManifestEvaluator {
 
     @Override
     public <T> Boolean ltEq(BoundReference<T> ref, Literal<T> lit) {
-      PositionAccessor positionAccessor = getPositionAccessorForFieldId(ref.fieldId());
-      ByteBuffer lowerBound = stats.get(positionAccessor.position()).lowerBound();
+      int pos = Accessors.toPosition(ref.accessor());
+      ByteBuffer lowerBound = stats.get(pos).lowerBound();
       if (lowerBound == null) {
         return ROWS_CANNOT_MATCH; // values are all null
       }
@@ -194,8 +181,8 @@ public class InclusiveManifestEvaluator {
 
     @Override
     public <T> Boolean gt(BoundReference<T> ref, Literal<T> lit) {
-      PositionAccessor positionAccessor = getPositionAccessorForFieldId(ref.fieldId());
-      ByteBuffer upperBound = stats.get(positionAccessor.position()).upperBound();
+      int pos = Accessors.toPosition(ref.accessor());
+      ByteBuffer upperBound = stats.get(pos).upperBound();
       if (upperBound == null) {
         return ROWS_CANNOT_MATCH; // values are all null
       }
@@ -212,8 +199,8 @@ public class InclusiveManifestEvaluator {
 
     @Override
     public <T> Boolean gtEq(BoundReference<T> ref, Literal<T> lit) {
-      PositionAccessor positionAccessor = getPositionAccessorForFieldId(ref.fieldId());
-      ByteBuffer upperBound = stats.get(positionAccessor.position()).upperBound();
+      int pos = Accessors.toPosition(ref.accessor());
+      ByteBuffer upperBound = stats.get(pos).upperBound();
       if (upperBound == null) {
         return ROWS_CANNOT_MATCH; // values are all null
       }
@@ -230,8 +217,8 @@ public class InclusiveManifestEvaluator {
 
     @Override
     public <T> Boolean eq(BoundReference<T> ref, Literal<T> lit) {
-      PositionAccessor positionAccessor = getPositionAccessorForFieldId(ref.fieldId());
-      PartitionFieldSummary fieldStats = stats.get(positionAccessor.position());
+      int pos = Accessors.toPosition(ref.accessor());
+      PartitionFieldSummary fieldStats = stats.get(pos);
       if (fieldStats.lowerBound() == null) {
         return ROWS_CANNOT_MATCH; // values are all null and literal cannot contain null
       }
